% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyNetwork.R
\name{sankeyNetworkOutput}
\alias{renderSankeyNetwork}
\alias{sankeyNetworkOutput}
\title{Title}
\usage{
sankeyNetworkOutput(outputId, width = "100\%", height = "500px")

renderSankeyNetwork(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{}

\item{width}{}

\item{height}{}
}
\description{
Title
}

