% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale-d3}
\alias{scale-d3}
\alias{scale_d3_aaas}
\alias{scale_d3_cosmic}
\alias{scale_d3_d3}
\alias{scale_d3_futurama}
\alias{scale_d3_gsea}
\alias{scale_d3_igv}
\alias{scale_d3_jama}
\alias{scale_d3_jco}
\alias{scale_d3_lancet}
\alias{scale_d3_locuszoom}
\alias{scale_d3_material}
\alias{scale_d3_nejm}
\alias{scale_d3_npg}
\alias{scale_d3_rickandmorty}
\alias{scale_d3_simpsons}
\alias{scale_d3_startrek}
\alias{scale_d3_tron}
\alias{scale_d3_uchicago}
\alias{scale_d3_ucscgb}
\title{Scientific Journal and Sci-Fi Themed Color Palettes for 'sankeyNetwork'
Scientific Journal and Sci-Fi Themed Color Palettes for 'sankeyNetwork'}
\usage{
scale_d3_aaas(palette = "default", alpha = 1, breaks = NULL)

scale_d3_cosmic(
  palette = c("hallmarks_light", "hallmarks_dark", "signature_substitutions"),
  alpha = 1,
  breaks = NULL
)

scale_d3_d3(
  palette = c("category10", "category20", "category20b", "category20c"),
  alpha = 1,
  breaks = NULL
)

scale_d3_futurama(palette = "planetexpress", alpha = 1, breaks = NULL)

scale_d3_gsea(palette = "default", alpha = 1, breaks = NULL)

scale_d3_igv(palette = c("default", "alternating"), alpha = 1, breaks = NULL)

scale_d3_jama(palette = "default", alpha = 1, breaks = NULL)

scale_d3_jco(palette = "default", alpha = 1, breaks = NULL)

scale_d3_lancet(palette = "lanonc", alpha = 1, breaks = NULL)

scale_d3_locuszoom(palette = "default", alpha = 1, breaks = NULL)

scale_d3_material(
  palette = c("red", "pink", "purple", "deep-purple", "indigo", "blue", "light-blue",
    "cyan", "teal", "green", "light-green", "lime", "yellow", "amber", "orange",
    "deep-orange", "brown", "grey", "blue-grey"),
  alpha = 1,
  breaks = NULL
)

scale_d3_nejm(palette = "default", alpha = 1, breaks = NULL)

scale_d3_npg(palette = "nrc", alpha = 1, breaks = NULL)

scale_d3_rickandmorty(palette = "schwifty", alpha = 1, breaks = NULL)

scale_d3_simpsons(palette = "springfield", alpha = 1, breaks = NULL)

scale_d3_startrek(palette = "uniform", alpha = 1, breaks = NULL)

scale_d3_tron(palette = "legacy", alpha = 1, breaks = NULL)

scale_d3_uchicago(
  palette = c("default", "light", "dark"),
  alpha = 1,
  breaks = NULL
)

scale_d3_uchicago(
  palette = c("default", "light", "dark"),
  alpha = 1,
  breaks = NULL
)

scale_d3_ucscgb(palette = "default", alpha = 1, breaks = NULL)
}
\arguments{
\item{palette}{Palette type. Details see \verb{pal_*} function in
\link[ggsci:ggsci-package]{ggsci}}

\item{alpha}{Transparency level, a real number in (0, 1].}

\item{breaks}{a set of data values to be mapped.}
}
\description{
Scientific Journal and Sci-Fi Themed Color Palettes for 'sankeyNetwork'
Scientific Journal and Sci-Fi Themed Color Palettes for 'sankeyNetwork'
}
