% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_d3_manual}
\alias{scale_d3_manual}
\title{Create your own d3 color scale
These functions allow you to specify your own set of mappings from data
values to aesthetic values.}
\usage{
scale_d3_manual(values, breaks = NULL)
}
\arguments{
\item{values}{a set of values to map data values to. The values will be
matched in order (usually alphabetical) with the levels of the data values,
or with breaks if provided. If this is a named vector, then the values will
be matched based on the names instead.}

\item{breaks}{a set of data values to be mapped.}
}
\description{
Create your own d3 color scale
These functions allow you to specify your own set of mappings from data
values to aesthetic values.
}
